Update the kasl-core and kasl-cranelift-backend

+++ src/runner/compiler/ptr_utils.rs
@@ -1,4 +1,4 @@
-use kasl::ast::scope_manager::BlueprintItem;
+use kasl_core::ast::scope_manager::BlueprintItem;
 use std::alloc::{Layout, alloc, dealloc};
 
 // --- OUTPUT & INPUT BUFFER ALLOCATION ---
@@ -10,7 +10,8 @@ pub(super) fn get_buffer_blueprint_ptr(
     let mut ptrs: Vec<*mut ()> = Vec::with_capacity(items.len());
     for item in items {
         let layout =
-            Layout::from_size_align(item.actual_size * buffer_size, item.align as usize).unwrap();
+            Layout::from_size_align(item.actual_size as usize * buffer_size, item.align as usize)
+                .unwrap();
         unsafe {
             let ptr = alloc(layout);
             if ptr.is_null() {
@@ -29,9 +30,11 @@ pub(super) fn deallocate_buffer_blueprint_ptr(
 ) {
     unsafe {
         for (item, ptr) in items.iter().zip(ptrs) {
-            let layout =
-                Layout::from_size_align(item.actual_size * buffer_size, item.align as usize)
-                    .unwrap();
+            let layout = Layout::from_size_align(
+                item.actual_size as usize * buffer_size,
+                item.align as usize,
+            )
+            .unwrap();
             if !ptr.is_null() {
                 dealloc(ptr as *mut u8, layout);
             }
@@ -44,7 +47,8 @@ pub(super) fn deallocate_buffer_blueprint_ptr(
 pub(super) fn get_blueprint_ptr(items: Vec<&BlueprintItem>) -> Vec<*mut ()> {
     let mut ptrs: Vec<*mut ()> = Vec::with_capacity(items.len());
     for item in items {
-        let layout = Layout::from_size_align(item.actual_size, item.align as usize).unwrap();
+        let layout =
+            Layout::from_size_align(item.actual_size as usize, item.align as usize).unwrap();
 
         unsafe {
             let ptr: *mut u8 = alloc(layout);
@@ -62,7 +66,8 @@ pub(super) fn get_blueprint_ptr(items: Vec<&BlueprintItem>) -> Vec<*mut ()> {
 
 pub(super) fn deallocate_blueprint_ptr(items: Vec<&BlueprintItem>, ptrs: Vec<*mut ()>) {
     for (item, ptr) in items.iter().zip(ptrs) {
-        let layout = Layout::from_size_align(item.actual_size, item.align as usize).unwrap();
+        let layout =
+            Layout::from_size_align(item.actual_size as usize, item.align as usize).unwrap();
 
         if !ptr.is_null() {
             unsafe {
